+++ src/components/Form/components/FormTable/FormTableHeader/FormTableHeader.component.tsx
@@ -1,17 +1,25 @@
 import { isString, type Optional } from "@ubloimmo/front-util";
+import styled from "styled-components";
 
 import { useFormContext } from "../../../Form.context";
 
-import { InputLabelText } from "@/components/InputLabel";
+import {
+  InputLabelText,
+  type InputLabelTextStyleProps,
+} from "@/components/InputLabel";
 import { Tooltip, type TooltipProps } from "@/components/Tooltip";
 import { FlexRowLayout, TableHeader, TableHeaderCell } from "@layouts";
 import { isEmptyString } from "@utils";
 
 import type { BuiltFormTableProps } from "../../../Form.types";
+import type { TextProps } from "@types";
 
-type FormTableHeaderProps = Pick<BuiltFormTableProps, "headers">;
+type FormTableHeaderProps = Pick<BuiltFormTableProps, "headers" | "colSpans">;
 
-export const FormTableHeader = ({ headers }: FormTableHeaderProps) => {
+export const FormTableHeader = ({
+  headers,
+  colSpans,
+}: FormTableHeaderProps) => {
   const { isEditing } = useFormContext();
   return (
     <TableHeader>
@@ -31,9 +39,9 @@ export const FormTableHeader = ({ headers }: FormTableHeaderProps) => {
           !isString(label) || isEmptyString(label) ? <>&nbsp;</> : label;
 
         return (
-          <TableHeaderCell key={key}>
+          <TableHeaderCell key={key} colSpan={colSpans[index] ?? 1}>
             <FlexRowLayout align="center" gap="s-2" justify={justify} fill>
-              <InputLabelText
+              <TableHeaderLabelText
                 color="primary-dark"
                 size="m"
                 weight="bold"
@@ -41,7 +49,7 @@ export const FormTableHeader = ({ headers }: FormTableHeaderProps) => {
                 $required={required && isEditing}
               >
                 {headerLabel}
-              </InputLabelText>
+              </TableHeaderLabelText>
               {headerTooltip && <Tooltip {...headerTooltip} />}
             </FlexRowLayout>
           </TableHeaderCell>
@@ -50,3 +58,10 @@ export const FormTableHeader = ({ headers }: FormTableHeaderProps) => {
     </TableHeader>
   );
 };
+
+const TableHeaderLabelText = styled(InputLabelText)<
+  InputLabelTextStyleProps & TextProps
+>`
+  white-space: nowrap;
+  min-width: max-content;
+`;