+++ src/layouts/Table/components/TableHeaderCell/TableHeaderCell.component.tsx
@@ -2,9 +2,10 @@ import styled from "styled-components";
 
 import { tableHeaderCellStyles } from "./TableHeaderCell.styles";
 
-import { useMergedProps } from "@utils";
+import { useClassName, useMergedProps, useTestId } from "@utils";
 
 import type { TableCellProps } from "../TableCell";
+import type { TestIdProps } from "@types";
 
 const defaultTableHeaderCellProps: TableCellProps = {
   children: null,
@@ -14,18 +15,21 @@ const defaultTableHeaderCellProps: TableCellProps = {
 /**
  * A table header cell component to label the corresponding column, used in `TableHeader`.
  *
- * @version 0.0.2
+ * @version 0.0.3
  *
  * @param {CellProps} props - The props for the component.
  * @return {JSX.Element} The rendered table header cell.
  */
-const TableHeaderCell = (props: TableCellProps): JSX.Element => {
+const TableHeaderCell = (props: TableCellProps & TestIdProps): JSX.Element => {
   const mergedProps = useMergedProps(defaultTableHeaderCellProps, props);
+  const testId = useTestId("table-header-cell", props);
+  const className = useClassName(props);
 
   return (
     <StyledTableHeaderCell
       colSpan={mergedProps.colSpan}
-      data-testid="table-header-cell"
+      data-testid={testId}
+      className={className}
     >
       <div data-testid="table-header-cell-inner">{props.children}</div>
     </StyledTableHeaderCell>